deployment/requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
seaborn
gdown
kaggle
